Add coverage for DateTime
